refactor: extract elements to files to better handle exported elements
