+++ autoequal/lib/src/public.dart
@@ -1,0 +1,10 @@
+import 'package:autoequal/src/src.dart';
+
+const Autoequal autoequal = Autoequal();
+
+@Deprecated('Use autoequal instead')
+const Autoequal autoequalMixin = Autoequal(mixin: true);
+
+const IgnoreAutoequal ignoreAutoequal = IgnoreAutoequal();
+
+const IncludeAutoequal includeAutoequal = IncludeAutoequal();
